miatoll: Add products without AOSP prefix
This is how Google builds the Pixel stock ROM, and it makes more sense
in my opinion. This is an Android device tree, and most of it is common
to all ROMs; rebranding the device for every custom ROM is a waste of
time when there are few differences.

Note that this has been added as a new product that overlays the AOSP
product to avoid merge conflicts in AOSP updates from Google and ease
commit exchanging with other ROMs based on AOSP trees.

Change-Id: I4825b3e8547e070604f52268dad52971190541d5
Signed-off-by: JamieHoSzeYui <henloboii@protonmail.ch>